Add configure options to retry failing scenarios
This change introduces a new option to retry all failed scenarios for the
specified number of times. By default they will not be retried.

It further expands the @flaky functionality by adding an option how often
these marked scenarios should be retried.

The configuration options can be set like this:

    Gurke.configure do |c|
      c.default_retries = 1
      c.flaky_retries = 3
    end